refactoring and badges
