+++ banking-utils/src/main/java/com/neoflex/java/util/CustomLogger.java
@@ -6,6 +6,7 @@ import lombok.extern.log4j.Log4j2;
  */
 @Log4j2
 public class CustomLogger {
+    private CustomLogger() {}
 
     /**
      * Функция вывода информации о вызванном методе и классе, из которого метод вызван
