+++ deal/src/test/java/com/neoflex/java/service/ConveyorAccessServiceTest.java
@@ -13,8 +13,8 @@ import com.neoflex.java.model.StatusHistoryElement;
 import com.neoflex.java.service.abstraction.ConveyorAccessService;
 import com.neoflex.java.service.abstraction.KafkaService;
 import com.neoflex.java.service.config.BaseTest;
-import com.neoflex.java.service.kafkaConfig.KafkaProducerConfig;
-import com.neoflex.java.service.kafkaConfig.KafkaTopicConfig;
+import com.neoflex.java.service.kafka_config.KafkaProducerConfig;
+import com.neoflex.java.service.kafka_config.KafkaTopicConfig;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
